+++ modules/init.mjs
@@ -18,9 +18,19 @@ Hooks.once('init', () => {
 Hooks.once('ready', () => {
   if(game.user.isGM && !window.pr.api.get('first-time-startup-notification-shown'))
     first_time_startup_notification()
+
+  ModuleSettings.add('toggle-actors-button-for-players', {
+    name: game.i18n.localize('FvttPartyResources.GMSettingsForm.ShowActorsButtonForPlayers'),
+    hint: game.i18n.localize('FvttPartyResources.GMSettingsForm.ShowActorsButtonForPlayersHint'),
+    default: true,
+    type: Boolean,
+    onChange: value => ActorDirectory.collection.render('actors')
+  });
 })
 
 Hooks.on('renderActorDirectory', (app, html, data) => {
+  if(!game.user.isGM && !ModuleSettings.get('toggle-actors-button-for-players')) return
+
   html
     .find(".directory-header")
     .prepend(`<div class="action-buttons flexrow"><button id="btn-dashboard"><i class="fas fa-calculator"> </i> ${game.i18n.localize('FvttPartyResources.Title')}</div>`)